update license; performance and compatibility improvements
Update date in license. Minor change to license text. Update contact info. Include intended URL of shader.

Instead of using an array of matrices to store the star location and orientation, use a switch statement to get the values for a given index. Fixes compilation problem for Safari on iOS, and improves performance for Intel iGPU.